schema_graph_critic/schema_miner/requirements.txt
PyGithub
requests
gitpython
python-dotenv
jsonschema
